Source object for rendering templates
This adds a source object for specifying and rendering template
files.  Create a source object for the given root of source files
you want to render.  It knows the root dir of the files it can
render plus the eruby/context class to render them with.

This is setup for implementing Deas' rendering API.