basic template engine with rendering api not implemented
This gets the basic template engine in place with the rendering
api not implemented for now.  This is setup for implementing the
render api at the source and hooking it up to Deas via the template
engine.

This covers the basics: setting the default logger/view locals,
creating the source object with the logger and setting up the render
API.