basic render implementation
This updates the source to add in basic rendering logic.  The source
builds a context class with any given locals, then loads to given
template file and evaluates it against the context.

This also updates the engine to define the render method and call
to the source to perform the render.  The implements the `render`
portion of the deas engine API.