render using a specific bufvar
This updates the source to render with a specific erb buffer variable.
This is setup for capture rendering as we will need to manipulate
the buffer to capture rendered output.  We need to know what var the
buffer is stored in so we can access it.

This also updates the `eruby` method to make it public and test it.
This should have been done previously with the meta-eruby method
changes as the meta methods are already public.  This also makes it
easier to test the bufvar logic.