implement new Deas `compile` api method
This implements the `compile` method from the updated Deas render API.
This method takes raw content and renders it against the given locals.

This allows deas-erubis to be used as a downstream engine when rendering
a template with multiple engines.  It is difficult to envision a
scenario where deas-erubis wouldn't be the top-level engine, but
this in theory supports it.  In most cases, however, deas-erubis
should be used as the top-level engine that renders to other downstream
template markups like Markdown, Scss, Coffee etc.

Note: this also does some cleanups on the source tests.  There was
a poorly styled context and the render no cache context was setup
incorrectly.