render with a given content block
This updates the render API to take an optional content block and
support yielding to it.  This implements layouts rendering.

Note: this removes the comments about looking at the handler and
handling any layouts.  Deas is going to do this.  This is a better
approach both because it centralizes the logic of building layout
content blocks and also each engine doesn't have to reimplement it.
Engines that don't want/need to support layouts can just ignore any
given content blocks.