add `partial` and `capture_partial` template helpers
This adds template helpers for rendering partials from other templates.
This covers both standard partials and partials that yield to given
content blocks.

This also removes support for an engine `capture_partial` method and
updates `partial` to take an optional content block (similar to the
`render` method.  This change makes `capture_partial` unnecessary.

Note: the tests need to be cleaned up a bit (which I'll do in a
coming PR).  They are doing way too many "system" tests in the unit
tests.  Plus having to add the `DeasSourceSpy` is not ideal.  Once
a new Deas rolls with the proper render API, I'll bring it in and
use its source directly and split out the system tests from the
unit tests.