system tests; test cleanups
This breaks out the unit tests that were rendering templates and
reworks them as system tests.  These should have been organized this
way originally.  This also switches to using an actual deas template
source object instead of a spy when testing the template helpers.

Finally this brings in the latest Deas to make sure we are current
with the latest render API.