update template helpers to take explicit arguments
Before we were just lazily passing on whatever was given.  However,
this fails if called with no locals as the deas source expects locals
to be passed.

This updates the template helpers to default the locals to `{}` if
none are specified.  This was missed in the original tests as no
templates called the helpers without locals.  This updates the tests
to cover this case.