use much-plugin on template helpers mixin
This brings in much-plugin and uses it on the template helpers mixin.
This ensures that the template helpers include logic only runs once.
This also updates some of the gem meta to get it up to our latest
standards/conventions.

Note: there were no template helper unit tests previously so this
adds some to test bringing in much-plugin.