source files: allow other exts between the name and .erb
The goal here is to still be able to look up source files when
they have multiple engine extensions (not just a trailing `.erb`).
For example, given the source file `template.html.md.erb`, deas-erubis
should be able to render that template with `render 'template.html'`.

This is prep for running templates through multiple engines in Deas.