(hotfix) rename `file_name` to `template_name`
This is a more appropriate name as this value can include path
info relative to the source root, not just a single file's name.
Plus, this is more consistent with the naming Deas uses.

This is just a cleanup to prep for changing how source files are
globbed to support multi-engine extensions.  I didn't want this
"noise" in that coming PR.