update source file globbing to support multi-engine extensions
Now that deas supports multi-engine extensions per template file,
deas-erubis can't know that every template file it is asked to render
will end in a `.erb` extension.

In addition, deas-erubis shouldn't be in the business of requiring
its source files to end in `.erb` anyway. If the developer has
configured deas-erubis on a custom extension, `.erubis` for example,
then deas-erubis shouldn't care.  If deas has asked deas-erubis to
render this named template, it should just do it.