source: require named templates end in any given extension
If an optional extension is specified, ensure that any template files
end in the given extension when rendering from the source file.  This
is more strict and better enforces the intent of specifying an extension.
This also allows you to specify fully qualified template names and
have them be recognized as it now only adds the ext to the glob if
needed.

Also, this updates the error message that is raised when a named
template source file can't be found.  The goal is to be very explicit
about what the source is looking for that it couldn't find and also
only adds extension info if an extension has been specified.