hotfix: update to use unique partial names - prep for Deas changes
This is prep for changes to how deas evaluates template names.  Names
have to be unique and non-composed.  If a name cannot uniquely identify
a template, Deas will now complain.  This is prep for that change -
the "_partial" name was composed in two different templates so this
switches one to be `'_partial'`.