Pass deas template engine ext to the source
This updates the template engine to pass its ext to the source. The
template engine ext is set by deas when the engine is registered.
This makes it so the source will use templates with the extension that
was used when registering the engine with deas. This is the expected
behavior and avoids confusion with the source randomly using a template
that does not match the registered extension.

This is part of fixing an issue with deas not rendering the expected
source template. This is because the source now no longer forces
template extensions to be `erb` (see PR 27). By not expecting an extension,
the source can't pick the correct file if there are multiple files with
the same composed name except for their extensions. For example, a common
pattern is to name a handler file and the template the same only the
handler will have `.rb` for its extension and the template would use
`.erb`. In this scenario, it's possible for the source to try and use
the ruby file as its template.