Use `File.read` when reading template files to use default encoding
This fixes an incompatible encodings error that can occur with the
current usage of `File.binread`. Using `File.binread` returns a
string with the ASCII-8BIT encoding. This can cause an
incompatible encoding error when a partial uses UTF-8 characters
and can't be coerced into an ASCII-8BIT string. What occurs is
that erubis builds a string of ruby that it instance eval. The
string starts out something like `@bufvar = ''` which creates a
string in the default encoding (for me UTF-8). This causes erubis
to take a UTF-8 string and then append (`<<`) on a ASCII-8BIT
string. As long as the template is both ASCII-8BIT and UTF-8
compatible this works but if the template uses UTF-8 characters
but was read with the ASCII-8BIT it will error.

This fixes the issue by switching to using `File.read` which
reads the template in the default encoding. This makes it so the
empty string that erubis creates and the templates are using the
same encoding and allows the user to control the encoding of
both via ruby's default internal encoding if desired. This seemed
to be the best option with not (easily) being able to effect the
encoding the erubis string uses.