Rename the testing `yield` template to `yield_no_view`
This renames the `yield` template in that's used in the tests
to instead be called `yield_no_view`. This was randomly erroring
because if you render a template called `"yield"` it will match
both the `yield.erb` and `yield_view.erb` and if it gets the
latter the tests error. By giving it the name `yield_no_view.erb`
it always chooses the correct template and the tests don't
randomly error.